Add CSS, minor changes throughout
CSS replaces the absurd amount of tags throughout the Document. I also improved the readabillity of the HTML.

I changed the bars to a border on the bottom, quicker and easier to use. Just do <h5 class = "block"> and it's there.

I also did some minor edits to the text to make it more gramatically correct.